+++ src/commands/index.ts
@@ -12,7 +12,6 @@ import pull from "./pull";
 import push from "./push";
 import repo from "./repo";
 import slice from "./slice";
-import starter from "./starter";
 import status from "./status";
 import sync from "./sync";
 import token from "./token";
@@ -70,10 +69,6 @@ export default createCommandRouter({
 		repo: {
 			handler: repo,
 			description: "Manage repositories",
-		},
-		starter: {
-			handler: starter,
-			description: "Download official starters",
 		},
 		type: {
 			handler: type_,