+++ src/commands/pull.ts
@@ -4,8 +4,8 @@ import { CommandError, createCommand, type CommandConfig } from "../lib/command"
 import { diffArrays } from "../lib/diff";
 import { getDirtyPaths, getGitRoot } from "../lib/git";
 import { getCustomTypes, getSlices } from "../lib/prismic/clients/custom-types";
-import { completeOnboardingStepsSilently } from "../lib/prismic/clients/repository";
 import { canonicalizeCustomType, canonicalizeSlice } from "../lib/prismic/models";
+import { completeOnboardingSteps } from "../lib/prismic/onboarding";
 import { isDescendant, relativePathname } from "../lib/url";
 import { findProjectRoot, getRepositoryName } from "../project";
 
@@ -88,7 +88,8 @@ export default createCommand(config, async ({ values }) => {
 		localCustomTypes.map((customType) => customType.model),
 		{
 			getKey: (model) => model.id,
-			equals: (remote, local) => JSON.stringify(canonicalizeCustomType(remote)) === JSON.stringify(local),
+			equals: (remote, local) =>
+				JSON.stringify(canonicalizeCustomType(remote)) === JSON.stringify(local),
 		},
 	);
 	const sliceOps = diffArrays(
@@ -96,7 +97,8 @@ export default createCommand(config, async ({ values }) => {
 		localSlices.map((slice) => slice.model),
 		{
 			getKey: (model) => model.id,
-			equals: (remote, local) => JSON.stringify(canonicalizeSlice(remote)) === JSON.stringify(local),
+			equals: (remote, local) =>
+				JSON.stringify(canonicalizeSlice(remote)) === JSON.stringify(local),
 		},
 	);
 
@@ -143,12 +145,11 @@ export default createCommand(config, async ({ values }) => {
 
 	await adapter.generateTypes();
 
-	await completeOnboardingStepsSilently({
+	await completeOnboardingSteps(["connectPrismic"], {
 		repo: await getRepositoryName(),
 		token,
 		host,
-		stepIds: ["connectPrismic"],
-	});
+	}).catch(() => {});
 
 	const totalTypes = customTypeOps.insert.length + customTypeOps.update.length;
 	const totalSlices = sliceOps.insert.length + sliceOps.update.length;